+++ src/main/java/com/google/cloud/tools/app/api/service/TrafficSplitConfiguration.java
@@ -20,7 +20,7 @@ import java.util.Map;
 
 /**
  * Identifies the traffic split for the version(s) of a given service or services
- * in {@link AppEngineService}.
+ * in {@link AppEngineServices}.
  */
 public interface TrafficSplitConfiguration extends ServiceSelectionConfiguration {
